+++ src/lib/functions/try/execute/getOrElseThrow.ts
@@ -2,6 +2,7 @@ import {Step, Try} from "../../../Try";
 import {Result} from "../../../Result";
 import {runSteps} from "../helpers";
 
+/** Execute and return value; when failed, throw error produced from original error. */
 export async function getOrElseThrow(tryObject: Try<unknown>, func: (err: Error) => Promise<Error> | Error){
     if(tryObject.$internal.isComputed){
         if(tryObject.$internal.finalResult?.isError())