+++ src/lib/functions/try/initializers/success.ts
@@ -1,6 +1,7 @@
 import {Result} from "../../../Result";
 
 
+/** Create a successful Result with the given value. */
 export async function success(value: any): Promise<Result> {
     return new Result().setValue(value);
 }